fix(zendesk-articles): fix pagination
As per the Zendesk API docs, the `next_page` field
is a URL that includes the `page` query parameter.

Which means that the params `{page: nextPage}` would
never work. I believe that instead, the initial idea
was to use the `currentPage` variable that is incremented
on each iteration of the loop.

In this commit, I'm fixing that logic and get rid
of the `nextPage` variable that is no longer needed.

[1]: https://developer.zendesk.com/api-reference/introduction/pagination/#using-offset-pagination