feat(notion): configure Notion database
We decided to take an opinionated approach and
manage the Notion database properties through this script.

We assume the database is dedicated to this script.

The default `Name` and `Tags` properties are removed.

All _managed fields_ are prefixed with `dust.`.
That way, if users want to add more fields, they can
do so without worrying about conflicts with the script.