feat(notion): add soft delete option
We were initially deleting the pages from Notion
when an assistant was deleted in Dust.

However, we have realized that there are no triggers
in Notion automation when a page is deleted.

Therefore, this commit introduces a new environment
variable `NOTION_SOFT_DELETE` which, when set to
`true`, will mark the `dust.status` property as
`deleted` instead of deleting the page.

That would allow new workflows where we could notify
a given slack channel when an assistant is deleted
in Dust.